StylizedFacts

Documentation for StylizedFacts.

StylizedFacts.Hurst_RS

StylizedFacts.Hurst_confint

StylizedFacts.Hurst_confint

StylizedFacts.absence_lin_autocor_test

StylizedFacts.compute_return_array

StylizedFacts.cond_heavy_tails_test

StylizedFacts.crosscor_volume_volatility_test

StylizedFacts.heavy_tails_test

StylizedFacts.leverage_effect_test

StylizedFacts.longmemory_volume_test

StylizedFacts.nonlinear_dependence_test

StylizedFacts.univariate_test_summary

StylizedFacts.volatility_clustering_test

StylizedFacts.Hurst_RS — Method

Hurst_RS(Z::AbstractVector) -> Tuple{Float64,Float64,Float64}

Estimate the Hurst exponent via rescaled range analysis.

Arguments

Z::AbstractVector: the real-valued time series vector

Keywords

adjust::Bool = true: correct the estimate via the Anis-Lloyd adjustment

Returns

Tuple{Float64,Float64,Float64}: the Hurst exponent in the first position, the log(n) values in the second position, and the log(R/S) values in the third position.

Throws

ArgumentError: throws an error if the length of Z is less than 500.

Notes

For several different lengths n, divide the time series Z` of length N into d subseries of length n, where n is an integral divisor of N. For each subseries m = 1, 2, ..., d, follow the steps presented in CITE. per confint paper -> log2, no n < 50, at least 512 observations

References

https://watermark.silverchair.com/63-1-111.pdf?token=AQECAHi208BE49Ooan9kkhWErcy7Dm3ZL9Cf3qfKAc485ysgAAAvUwggLxBgkqhkiG9w0BBwagggLiMIIC3gIBADCCAtcGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMZn9ZfD4HT8p48oH7AgEQgIICqFFc432q0RZTSB95JYv8I7i-nrSWUQdoHy2CfmhMYjbyoUXNCkoCV1HuMbJC28IYbaz0Li3n2PcmUeZjBkqoqecOZLQMypFp5PCR2xxqRqOUW3VQP2lrvPH-pJe9znb-RBrgY2wNF0zkvxadG1f7GAntRh2JBchmRWoCLKANBerIZnzKlaufIpRMOAq6uajOo36gzoNvlBGI2WghtJhsZuQqCFwHYRcTtmefCGIJRr4xOXs7D1ap9F5OwL8NhsRuZ8pPq0YQuQctiSjU2DRmyD0nJI1BybRb-678jD06sgI0XH7HuCWoUAVSqYHjuX1gvdYac1LHZ3nIq4AsNmVSPJzr3cZ9-C4rwJ-kmbZCjO2TXavD28xpLcwFuOzhjrwoQGLfhLFvYiyjJWLyUMnH4sFDS2QKR39IVX37qpN6B3xipuowiAeY13dyzWwzE3TjIGBumESD5UHOu3Q05yMi7DHlGZq-R9G2SxYJB4PnOOqgWvR7uMgCGw1Xjg0jWs9UrpTcDuQMTe0QoKjWUwjNTI5sodgI8YBogBCCXankFtuqYSsiElfl5CI7CcWj33VX6H7sMv2zSE8j-XyuOZKGHA5VnvF4N2mjBupHbId7rh6olQ8ANub2jqKsv8colB2Zv6ygtVq0INA760Q3oduR2fl1TAc5Y-VxJ17GtWUjRPLkbgsBhITu6KpqRdQZIZnQIl7dUGh9TdLtxcuMkGEkLqSk3qsCQ02Vj7zhb5zwdk4L3dZ7UMU8Ov1YlUJExbkuRnDKJfTWi9FOkDnYTbapjLbsWsbW4hJA_T1es65nEOfQVs2IaXu2Sbg6QLMiqHgPnVyauzMVFMhPtm-j2Coj0dc4zLFtRDct9aUxX5NNhrThVjwJurtXh9MhcKtWyAFMdq5u https://arxiv.org/pdf/1805.08931.pdf https://arxiv.org/pdf/cond-mat/0103510.pdf

StylizedFacts.Hurst_confint — Method

Hurst_confint(N::Int; level::Symbol=:ninetyfive) -> Tuple{Float64,Float64}

Returns the confidence interval for Hurst exponents estimated via rescaled range analysis.

Arguments

N::Int: the length of the time series

Keywords

level::Symbol=:ninetyfive: the confidence level (options -> :ninety, :ninetyfive (default), :ninetynine)

Returns

Tuple{Float64,Float64,Float64}: the lower bound of the Hurst exponent in the first position and the upper bound of the Hurst exponent in the second position (according to the confidence interval).

References

https://arxiv.org/pdf/cond-mat/0103510.pdf

StylizedFacts.Hurst_confint — Method

Hurst_confint(Z::AbstractVector; level::Symbol=:ninetyfive) -> Tuple{Float64,Float64}

Returns the confidence interval for Hurst exponents estimated via rescaled range analysis.

Returns a Tuple with the lower and upper bound of the Hurst exponent (according to the confidence interval).

Arguments

Z::AbstractVector: the real-valued time series vector

Keywords

level::Symbol=:ninetyfive: the confidence level (options -> :ninety, :ninetyfive (default), :ninetynine)

Returns

Tuple{Float64,Float64,Float64}: the lower bound of the Hurst exponent in the first position and the upper bound of the Hurst exponent in the second position (according to the confidence interval).

Throws

ArgumentError: throws an error if the length of Z is less than 500.

References

https://arxiv.org/pdf/cond-mat/0103510.pdf

StylizedFacts.absence_lin_autocor_test — Method

absence_lin_autocor_test(r::AbstractVector) -> Bool

Compute the autocorrelation function and estimate the Hurst exponent to test for the absence of linear autocorrelation in the real-valued vector of returns r.

Arguments

r::AbstractVector{T}: the real-valued vector of returns

Returns

Bool: true if the sample exhibits the absence of linear autocorrelation and false otherwise.

Notes

The test...

References

StylizedFacts.compute_return_array — Method

compute_return_array(data::DataFrame; cc=true) -> DataFrame

Compute the return array of the price series DataFrame

Arguments

data::DataFrame: the price series DataFrame

Keywords

cc::Bool=true: denotes the use of close-to-close returns series (set cc=false if open-to-close returns series is desired)

Returns

DataFrame: the first column :timestamp denotes the timestamp and the second column

:R denotes the return instance corresponding to the timestamp

StylizedFacts.cond_heavy_tails_test — Method

cond_heavy_tails_test(r::Vector) -> Bool

Compute the kurtosis and Jarque-Bera statistic of the volatility-corrected residual time series to test for the presence of conditional heavy tails in the real-valued vector of returns r.

Arguments

r::AbstractVector{T}: the real-valued vector of returns

Returns

Bool: true if the sample exhibits conditional heavy tails and false otherwise.

Notes

The test...

References

StylizedFacts.crosscor_volume_volatility_test — Method

crosscor_volume_volatility_test(u::Vector, r::Vector)

Compute the cross-correlation function to test for the presence of volatility clustering in the real-valued vector of volume u and real-valued vector of returns r.

Arguments

u::AbstractVector{T}: the real-valued vector of volume

r::AbstractVector{T}: the real-valued vector of returns

Returns

Bool: true if the sample exhibits volatility clustering and false otherwise.

Notes

The test...

References

StylizedFacts.heavy_tails_test — Method

heavy_tails_test(r::AbstractVector) -> Bool

Compute the kurtosis and Jarque-Bera statistic to test for the presence of heavy tails in the real-valued vector of returns r.

Arguments

r::AbstractVector{T}: the real-valued vector of returns

Returns

Bool: true if the sample exhibits heavy tails and false otherwise.

References

StylizedFacts.leverage_effect_test — Method

leverage_effect_test(r::Vector) -> Bool

Compute the cross-correlation of returns with subsequent squared returns to test for the presence of the leverage effect in the real-valued vector of returns y.

Arguments

r::AbstractVector{T}: the real-valued vector of returns

Returns

Bool: true if the sample exhibits the leverage effect and false otherwise.

Notes

The test...

References

StylizedFacts.longmemory_volume_test — Method

longmemory_volume_test(u::Vector) -> Bool

Compute the autocorrelation function and estimate the Hurst exponent to test for the presence of long memory in the real-valued vector of volume u.

Arguments

u::AbstractVector{T}: the real-valued vector of volume

Returns

Bool: true if the sample exhibits long memory and false otherwise.

Notes

The test...

References

StylizedFacts.nonlinear_dependence_test — Method

nonlinear_dependence_test(r::Vector) -> Bool

Compute nonlinear transforms to test for the presence of nonlinear dependence in the real-valued vector of returns r.

Arguments

r::AbstractVector{T}: the real-valued vector of returns

Returns

Bool: true if the sample exhibits nonlinear dependence and false otherwise.

Notes

The test...

References

StylizedFacts.univariate_test_summary — Method

univariate_test_summary(ticker_symbol_array::Vector{String},
                        price_data_dictionary::Dict{String,DataFrame},
                        return_data_dictionary::Dict{String,DataFrame};
                        cc::Bool=true,
                        print_msg::Bool=true)

Compute the result of each univariate stylized facts test for each ticker in ticker_symbol_array.

Arguments

ticker_symbol_array::Vector{String}: list of ticker symbols to be tested

price_data_dictionary::Dict{String,DataFrame}: dictionary with a ticker symbol key and corresponding value of a DataFrame with OHLC price and volume data

return_data_dictionary::Dict{String,DataFrame}: dictionary with a ticker symbol key and corresponding value of a DataFrame with returns series data (see: computereturnarray() method)

Keywords

cc::Bool=true: denotes the use of close-to-close returns series (set cc=false if open-to-close returns series)

print_msg::Bool=true: print out a summary of the stylized fact tests results

Returns

Tuple{DataFrame,DataFrame}: the first DataFrame stores the result true if the ticker exhibits the stylized fact for that respective test column and false otherwise; the second DataFrame is a summary of the percentage of tickers that were able to reproduce the stylized fact of each respective test column.

References

StylizedFacts.volatility_clustering_test — Method

volatility_clustering_test(r::AbstractVector) -> Bool

Compute the autocorrelation function and estimate the Hurst exponent to test for the presence of volatility clustering in the real-valued vector of returns r.

Arguments

r::AbstractVector{T}: the real-valued vector of returns

Returns

Bool: true if the sample exhibits volatility clustering and false otherwise.

Notes

The test...

References